Cleanup BigDipper API and types
