+++ src/types/bigDipper.ts
@@ -1,39 +1,3 @@
-import { Coin } from './node';
-
-export class Account {
-  public accountBalance: { coins: Coin[] };
-  public delegationBalance: { coins: Coin[] };
-  public unbondingBalance: { coins: Coin[] };
-  public rewardBalance: [{ coins: Coin[] }];
-  public vesting_account: {
-    id: string;
-    type: string;
-    original_vesting: Coin[];
-    start_time: number;
-    end_time: number;
-  }[];
-
-  constructor(
-    account_balance: { coins: Coin[] },
-    delegation_balance: { coins: Coin[] },
-    unbonding_balance: { coins: Coin[] },
-    reward_balance: [{ coins: Coin[] }],
-    vesting_account: {
-      id: string;
-      type: string;
-      original_vesting: Coin[];
-      start_time: number;
-      end_time: number;
-    }[]
-  ) {
-    this.accountBalance = account_balance;
-    this.delegationBalance = delegation_balance;
-    this.unbondingBalance = unbonding_balance;
-    this.rewardBalance = reward_balance;
-    this.vesting_account = vesting_account;
-  }
-}
-
 export interface TotalSupplyResponse {
   supply: [
     {
@@ -43,6 +7,29 @@ export interface TotalSupplyResponse {
           amount: string;
         }
       ];
+    }
+  ];
+}
+
+export interface TotalStakedCoinsResponse {
+  staking_pool: [
+    {
+      bonded_tokens: string;
+    }
+  ];
+}
+
+export interface ActiveValidatorsResponse {
+  validator_info: [
+    {
+      operator_address: string;
+      validator: {
+        validator_voting_powers: [
+          {
+            voting_power: number;
+          }
+        ];
+      };
     }
   ];
 }
