+++ src/types/node.ts
@@ -70,29 +70,6 @@ export interface ValidatorDetailResponse {
   };
 }
 
-export interface ActiveValidatorsResponse {
-  validator_info: [
-    {
-      operator_address: string;
-      validator: {
-        validator_voting_powers: [
-          {
-            voting_power: number;
-          }
-        ];
-      };
-    }
-  ];
-}
-
-export interface TotalStakedCoinsResponse {
-  staking_pool: [
-    {
-      bonded_tokens: string;
-    }
-  ];
-}
-
 export interface AccountBalanceInfos {
   totalBalance: number;
   availableBalance: number;
